fft_eval: Implement FSFE REUSE 3.0 specification
Not all files in the FFT_eval repository had easily detectable copyright
attached to it. This makes it harder for other persons to distribute the
files or use the files in different context.

The REUSE 3.0 specification works around by this using various ways to save
the copyright information and at the same time provides tools to check this
& export the bill of material for the project.

  # check copyright information
  reuse lint

  # create bill of material in SPDX format
  reuse spdx

Signed-off-by: Sven Eckelmann <sven@narfation.org>